refactor(state-router): add exports directory
