+++ packages/@sanity/state-router/package.json
@@ -2,16 +2,13 @@
   "name": "@sanity/state-router",
   "version": "2.18.0",
   "description": "A path pattern => state object bidirectional mapper",
-  "main": "./lib/index.js",
-  "types": "./dist/dts",
+  "main": "./lib/_exports/index.js",
+  "types": "index.d.ts",
   "//": "the typesVersion config below is a workaround for TypeScript's lack of support for package exports",
   "typesVersions": {
     "*": {
-      "dist/dts/index.d.ts": [
-        "dist/dts/index.d.ts"
-      ],
-      "components": [
-        "dist/dts/components/index.d.ts"
+      "*": [
+        "dist/dts/_exports/*"
       ]
     }
   },
