--- packages/@sanity/state-router/src/index.ts
@@ -1,14 +1,0 @@
-import resolveStateFromPath from './resolveStateFromPath'
-import resolvePathFromState from './resolvePathFromState'
-
-export type {Router} from './types'
-export {default as route} from './route'
-export {RouterContext} from './RouterContext'
-
-export * from './useRouter'
-export * from './useRouterState'
-
-export {resolveStateFromPath}
-export {resolvePathFromState}
-
-export type {HOCRouter} from './components/types'
